chore: add production builds
