+++ package.json
@@ -34,8 +34,8 @@
     "url": "https://github.com/wexond/wexond/issues"
   },
   "scripts": {
-    "dev": "concurrently \"webpack-dev-server --config webpack.config.dev.js --mode development\" \"webpack --mode development\"",
-    "build": "node fuse.js",
+    "dev": "concurrently \"webpack-dev-server --config webpack.config.renderer.dev.js\" \"webpack\"",
+    "build": "concurrently \"webpack --config webpack.config.renderer.prod.js\" \"webpack --config webpack.config.prod.js\"",
     "start": "cross-env ENV='dev' electron .",
     "compile-win32": "npm run build && electron-builder -w -p always",
     "compile-darwin": "npm run build && electron-builder -m -p always",
@@ -53,8 +53,6 @@
     "@types/axios": "0.14.0",
     "@types/chrome": "0.0.87",
     "@types/gsap": "1.20.2",
-    "@types/leveldown": "4.0.0",
-    "@types/levelup": "3.1.1",
     "@types/nedb": "1.8.8",
     "@types/node": "12.6.9",
     "@types/react": "16.8.24",
@@ -66,7 +64,6 @@
     "cross-env": "5.2.0",
     "electron": "7.0.0-beta.1",
     "electron-builder": "21.2.0",
-    "electron-extensions": "2.0.2",
     "electron-react-devtools": "0.5.3",
     "electron-updater": "4.1.2",
     "file-type": "12.1.0",
@@ -103,6 +100,7 @@
     "extract-file-icon": "0.3.1",
     "mouse-hooks": "0.4.1",
     "node-vibrant": "3.2.0-alpha",
-    "node-window-manager": "1.0.4"
+    "node-window-manager": "1.0.4",
+    "electron-extensions": "2.0.2"
   }
 }
